refactor: move mediation record retrieval util
